# src/backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: RentalFlow API
  description: API documentation for the RentalFlow microservices ecosystem.
  version: 1.0.0
servers:
  - url: https://rentalflow.onrender.com
    description: Production Server (API Gateway)
  - url: http://localhost:8080
    description: Local Development Server (API Gateway)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/payments/initialize:
    post:
      summary: Start a payment flow
      tags: [Payment]
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url: { type: string }
...